% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitwise.R
\name{bitwise.integer2double}
\alias{bitwise.integer2double}
\title{Bitwise value conversion}
\usage{
bitwise.integer2double(ival)
}
\arguments{
\item{ival}{integer values, could be integer, double, logical type,
but must be integer values.}

\item{dval}{double values of double type.}
}
\description{
Reinterpret a value of one type as a bitwise identical value of
another type. After reinterpretion, the values are different, but
the underlying bits representation with respect to the two types
are the same.
}
\details{
These functions mimic the \code{reinterpret_cast<A&>(val)}
functionality of C++. Actually, they are implemented using C++
\code{reinterpret_cast}.
}
