% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reader.R
\docType{class}
\name{Reader}
\alias{Reader}
\title{csv file reader which can read the file sequentially}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Reader
}
\description{
csv file reader which can read the file sequentially
}
\keyword{data}
